+++ src/MetaheuristicsPlatform/Algorithms/PSO/Scientific/InertiaWeightPsoReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.PSO.Scientific;
+
+public static class InertiaWeightPsoReferences
+{
+    public static ScientificReference ShiEberhart1998 { get; } =
+        new(
+            "Shi & Eberhart",
+            1998,
+            "A Modified Particle Swarm Optimizer",
+            "Proceedings of the 1998 IEEE International Conference on Evolutionary Computation, 69-73",
+            "10.1109/ICEC.1998.699146");
+}